creacion de model, rout y controller tabla favorites
