+++ server/server.js
@@ -6,6 +6,7 @@ import { logger } from "logger-express";
 import userRoutes from "./config/routes/userRoutes.js";
 import productRoutes from "./config/routes/productRoutes.js";
 import categoryRoutes from "./config/routes/categoryRoutes.js";
+import favoriteRoutes from "./config/routes/favoriteRoutes.js";
 
 
 const app = express();
@@ -20,5 +21,6 @@ app.use(logger());
 app.use("/api/v1", userRoutes);
 app.use("/api/v1", productRoutes);
 app.use("/api/v1", categoryRoutes);
+app.use("/api/v1", favoriteRoutes);
 
 app.listen(PORT, console.log(`¡Servidor encendido en el puerto! ${PORT}`));
